Move HeaderOptions
