+++ src/data/data-table/shared/DataTable.js
@@ -13,7 +13,6 @@ import {
   Link,
   RadioButton,
 } from 'src/components/common';
-import { HeaderOptions } from 'src/components/data/data-utils';
 import { icon } from 'src/components/icons';
 import { ConfirmedSearchInput } from 'src/components/input';
 import { MenuButton } from 'src/components/MenuButton';
@@ -44,6 +43,7 @@ import {
 } from '../entity-service/SavedColumnSettings';
 import { SingleEntityEditor } from '../entity-service/SingleEntityEditor';
 import { EditDataLink } from './EditDataLink';
+import { HeaderOptions } from './HeaderOptions';
 
 const entityMap = (entities) => {
   return _.fromPairs(_.map((e) => [e.name, e], entities));
